refactor: 모달 초기화 수정 및 컴포넌트 생성 파일 폴더로 분리
